+++ source/scan-tool/ZipHandler.hpp
@@ -35,6 +35,10 @@ namespace virustotal
 class ZipHandler: public Handler
 {
   public:
+    /** constructor */
+    ZipHandler(const bool ignoreErrors = false);
+
+
     /** \brief scan a given file using the implemented handling mechanism
      *
      * \param strategy  reference to the current scan strategy
@@ -65,6 +69,23 @@ class ZipHandler: public Handler
               std::unordered_map<std::string, std::string>& queued_scans,
               std::chrono::time_point<std::chrono::steady_clock>& lastQueuedScanTime,
               std::vector<std::pair<std::string, int64_t> >& largeFiles) override;
+
+
+    /** \brief checks whether or not this handler ignores extraction failures
+     *
+     * \return Returns true, if extraction failure does not produce an error or
+     * an exception. Returns false, if extraction failure causes errors.
+     */
+    bool ignoreExtractionErrors() const;
+
+
+    /** \brief sets whether or not extraction errors will be ignore
+     *
+     * \param ignore  whether or not to ignore extraction failure
+     */
+    void ignoreExtractionErrors(const bool ignore);
+  private:
+    bool m_IgnoreExtractionErrors; /**< whether to continue, if extraction fails */
 }; //class
 
 } //namespace